requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
openpyxl>=3.1.0
pydantic>=2.6.0
pyyaml>=6.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pymupdf>=1.23.0
google-api-python-client>=2.100.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
tenacity>=8.2.0
litellm>=1.78.0
openai>=1.99.5
jsonschema>=4.22.0
email-validator>=2.0.0
aiohttp>=3.10.0
pytest>=8.0.0
python-dotenv>=1.0.0
docling>=2.0.0
